refactor(core): split context.ts into focused modules
Extract the 1378-line monolith into:
- context/helpers.ts (291 lines): parsing, normalization, and utility functions
- context/context-manager.ts (1104 lines): ContextManager class
- context/index.ts: re-export shim
- context.ts: thin proxy re-export for backward compatibility

Closes #44
