+++ packages/core/src/context/index.ts
@@ -1,0 +1,1 @@
+export { ContextManager, createContextManager } from './context-manager.js';
